requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==4.4.2
emoji==2.7.0
exceptiongroup==1.1.2
fake-useragent==1.2.1
ffmpeg==1.4
frozenlist==1.4.0
gmail==0.6.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
kaitaistruct==0.10
moviepy==1.0.3
multidict==6.0.4
notifypy==1.0.3.0
numpy==1.25.2
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
playwright==1.36.0
proglog==0.1.10
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyee==9.0.4
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.11.2
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
TgCrypto==1.2.5
tiktok-uploader==1.0.1
TikTokApi==6.0.1
tiktokapipy==0.2.3
toml==0.10.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
twilio==8.5.0
typing_extensions==4.7.1
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
zstandard==0.21.0
